+++ api/authentication/urls.py
@@ -4,7 +4,7 @@ from allauth.socialaccount import providers
 
 from rest_framework_simplejwt.views import token_obtain_pair,token_refresh
 from django.contrib.auth import views as auth_views
-from .views import (
+from authentication.views import (
     UserAPIView,
     UsersAPIView,
     VerifyPassUserAPIView,
@@ -15,8 +15,8 @@ from .views import (
     TeacherAPIView,
     StudentAPIView,
     StudentsAPIView,
-    SendMailAPIView,
    )
+from utils.views import SendMailAPIView
 
 urlpatterns = [
     
@@ -25,18 +25,24 @@ urlpatterns = [
     path('users/<int:id>/',UserAPIView.as_view(), name='user'),
     path('users/token_obtain/', token_obtain_pair, name='login'),    
     path('users/inactive/', UserInactiveAPIView.as_view(), name='account_inactive'),
-    path('users/send_mail/', SendMailAPIView.as_view(), name='sendmail'),
     path('users/recover_pass/', RecoveryAPIView.as_view(), name='recover'),
     path('users/token_refresh/', token_refresh, name='refresh'),
     path('users/verify/<str:code>/', VerifyUserAPIView.as_view(), name='verify'),
     path('users/recovery/<str:code>/', VerifyPassUserAPIView.as_view(), name='completerecover'),
-    
+    path('users/send_mail/', SendMailAPIView.as_view(), name='send_mail'),
+    #teacher api
     path('teachers/', TeachersAPIView.as_view(), name='teahers'),
     path('teachers/<int:id>/', TeacherAPIView.as_view(), name='teacher'),
-    
+    #students api
     path('students/', StudentsAPIView.as_view(), name='students'),
     path('students/<int:id>/', StudentAPIView.as_view(), name='student'),
-    
+    #moderator api
+    #path('moderator/ban_user/',view.as_view(),name='ban_user'),
+    #path('moderator/mass_mail/',view.as_view(),name=''),
+    #path('moderator/set_news/',view.as_view(),name=''),
+    #admin api
+    #path('admin/set_moder/',view.as_view(),name=''),
+    #path('admin/user/',view.as_view(),name=''),
 ]
 
 # This is for social auth
