+++ api/ext_news/models.py
@@ -11,20 +11,4 @@ class News(models.Model):
     def __str__(self):
         return self.title
 
-    @classmethod
-    def mailing(self, data):
-        queryset = list(News.objects.all()[:3])
-        context = {
-                'user': data.email, 
-                'data': queryset
-                #'user': data,
-                #'title': title,
-                #'discription': discription,
-                #'news_link': news_link,
-                #'images_link': images_link
-                }
-        msg = EmailMessage(subject='Новини за тиждень',
-                body=render_to_string('ext_news/mail/mail.html',context), 
-                to=[data.email])
-        msg.content_subtype = 'html'
-        msg.send() 
+
